requirements.txt
asgiref==3.6.0
Django==3.2.18
django-ckeditor==6.3.2
django-ckeditor-5==0.2.4
django-js-asset==2.0.0
django-taggit==3.1.0
Pillow==9.4.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
